Move `RawModeConfig` to `config::raw` Module
